implement jwt token and project private api

+++ index.js
@@ -1,5 +1,7 @@
 const express = require('express');
 const cors = require('cors')
+const jwt = require('jsonwebtoken');
+const cookieParser = require('cookie-parser');
 const app = express();
 require('dotenv').config();
 const PORT = process.env.PORT || 3000;
@@ -12,7 +14,9 @@ app.use(cors({
   ],
   credentials: true
 }));
+
 app.use(express.json());
+app.use(cookieParser());
 
 const { MongoClient, ServerApiVersion, ObjectId } = require('mongodb');
 const uri = `mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASSWORD}@cluster0.jd7el.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0`;
@@ -43,6 +47,36 @@ async function connectToMongoDB() {
 
 connectToMongoDB();
 
+
+const verifyToken = (req, res, next) => {
+  const token = req?.cookies?.token;
+  if (!token) {
+    return res.status(401).send({ message: 'unAuthorized access' })
+  }
+
+  jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, decoded) => {
+    if (err) {
+      return res.status(401).send({ message: 'unauthorized access' })
+    }
+    req.user = decoded;
+    next();
+  })
+
+}
+
+// Auth related APIs
+app.post('/jwt', async (req, res) => {
+  const user = req.body;
+  const token = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET, { expiresIn: '1h' });
+
+  res
+    .cookie('token', token, {
+      httpOnly: true,
+      secure: false, //for localhost
+    })
+    .send({ success: true })
+
+});
 // Define routes
 app.get('/', (req, res) => {
   res.send('Welcome to the Express.js server!');
@@ -65,7 +99,7 @@ app.get('/assignments', async (req, res) => {
   }
 });
 
-app.post('/assignments', async (req, res) => {
+app.post('/assignments', verifyToken, async (req, res) => {
   try {
     const data = req.body;
     const result = await assignmentsCollection.insertOne(data);
@@ -74,7 +108,7 @@ app.post('/assignments', async (req, res) => {
     console.log(e)
   }
 })
-app.get("/assignments/:id", async (req, res) => {
+app.get("/assignments/:id", verifyToken, async (req, res) => {
   const { id } = req.params;
   try {
     const result = await assignmentsCollection.findOne({ _id: new ObjectId(id) })
@@ -84,10 +118,15 @@ app.get("/assignments/:id", async (req, res) => {
     res.status(500).json({ message: "Internal server error" });
   }
 })
-app.delete("/assignments/:id", async (req, res) => {
+app.delete("/assignments/:id", verifyToken, async (req, res) => {
   const { id } = req.params;
 
   try {
+    const data = await assignmentsCollection.findOne({ _id: new ObjectId(id) });
+    if (req.user.email !== data.email) {
+      return res.status(403).send({ message: 'forbidden access' });
+    }
+
     const result = await assignmentsCollection.deleteOne({ _id: new ObjectId(id) });
 
     if (result.deletedCount === 1) {
@@ -100,11 +139,15 @@ app.delete("/assignments/:id", async (req, res) => {
     res.status(500).json({ message: "Internal server error" });
   }
 });
-app.put('/assignments/:id', async (req, res) => {
+app.put('/assignments/:id', verifyToken, async (req, res) => {
   try {
-    const { id } = req.params; 
-    const updatedData = req.body; 
+    const { id } = req.params;
+    const updatedData = req.body;
 
+    const data = await assignmentsCollection.findOne({ _id: new ObjectId(id) });
+    if (req.user.email !== data.email) {
+      return res.status(403).send({ message: 'forbidden access' });
+    }
     const result = await assignmentsCollection.updateOne(
       { _id: new ObjectId(id) },
       { $set: updatedData }
@@ -122,12 +165,18 @@ app.put('/assignments/:id', async (req, res) => {
 });
 
 // ----------------------------submissions api---------------------
-app.get('/submissions', async (req, res) => {
+app.get('/submissions', verifyToken, async (req, res) => {
   const email = req.query?.email;
   const status = req.query?.status;
+
   const query = {};
-  if (email) query.userEmail = email;
-  if(status) query.status = status
+  if (email) {
+    if (req.user.email !== email) {
+      return res.status(403).send({ message: 'forbidden access' });
+    }
+    query.userEmail = email;
+  }
+  if (status) query.status = status
   try {
     const cursor = submissionsCollection.find(query);
     const result = await cursor.toArray();
@@ -137,7 +186,7 @@ app.get('/submissions', async (req, res) => {
     res.status(500).send('Error fetching assignments');
   }
 });
-app.post('/submissions', async (req, res) => {
+app.post('/submissions', verifyToken, async (req, res) => {
   try {
     const data = req.body;
     const result = await submissionsCollection.insertOne(data);
@@ -146,14 +195,14 @@ app.post('/submissions', async (req, res) => {
     console.log(e)
   }
 })
-app.patch('/submissions/:id', async (req, res) => {
+app.patch('/submissions/:id', verifyToken, async (req, res) => {
   const id = req.params.id;
   const data = req.body;
   const filter = { _id: new ObjectId(id) };
   const updatedDoc = {
     $set: {
       status: data.status,
-      obtainedMarks : data.obtainedMarks,
+      obtainedMarks: data.obtainedMarks,
       feedback: data.feedback
     }
   }
@@ -165,9 +214,3 @@ app.listen(PORT, () => {
   console.log(`Server is running on http://localhost:${PORT}`);
 });
 
-// Ensure the MongoDB client closes gracefully on server termination
-// process.on('SIGINT', async () => {
-//   console.log('Closing MongoDB connection...');
-//   await client.close();
-//   process.exit(0);
-// });
